+++ changelog.md
@@ -2,6 +2,7 @@ InsaneLib's Changelog
 
 ## 1.8.1
 * Added `ILItemTier`
+* Added `ILMobEffect`
 * Added `Feature#getConfigOption` (and renamed `Feature#setConfig` to `Feature#setConfigOption`)
 * Fixed nested Blacklist and MinMax configs not popping the path correctly