+++ changelog.md
@@ -7,6 +7,7 @@ InsaneLib's Changelog
 * Fixed nested Blacklist and MinMax configs not popping the path correctly
 * Fixed `WeightedRandom` still using `Random` instead of `RandomSource
 * Fixed 'Fix Follow Range' not working
+* Fixed `@Config` annotation having the wrong min value by default
 
 ## 1.8.0
 * Bump version for Minecraft 1.19.3. 1.8.x+ is now 1.19.3 and 1.7.x it's 1.19.2